Moving time/date utilities into utilities/temporal.py
